Move idl tests into separate files
